Bug 1397801 - In VisitCallExpr, wait until we get the real location of the callee to decide if it is interesting.
The start of a declaration of a thing being called can end up being
the return type, such as bool. When it is a primitive type like bool,
then the normalized location ends up being uninteresting, so nothing
ends up being reported for the call.

(billm, mrbkap and I worked together to figure this out.)