Bug 1415700, part 1 - Don't duplicate file paths in crossref.rs.
Rather than making many copies of the file names, create a single list
of them before starting the loop, then use references to them. The
explicit type coercion is needed.

This changes the behavior in the "bad line number" case because it
will now keep going on the rest of the data, rather then stopping
immediately.