Bug 1415700, part 3 - Only process each line in the file once.
Do the trimming of each line before we start looking at each
line. This should keep it from being done for each line repeatedly,
though I saw no speed improvement. I think it should reduce peak
memory usage because we don't need to keep around the unprocessed
lines any more. On the down side, we'll end up running trim on lines
that are never referred to. I'm not entirely sure this is an
improvement, but I think it is a little cleaner.