Bug 1416876 - Make crossref output stable.
One file is generated by iterating over a HashMap keyed by heap
allocated structures (table) so the output is not consistent. Using a
BtreeMap fixes this.

id_table technically has this problem, too, though in practice because
it is keyed with strings the order seems to be consistent.

The third output is generated from a vector, so it should be okay.