Make the blame repo optional.
This makes it so that we can specify a git repo in the config.json
without a blame repo, and get the partial functionality that unlocks.
In particular this enables the navigation panel if the tree has a git
repo specified, even if it doesn't have the blame repo.
This also adds the searchfox git repo to the tests/config.json file, so
now using the test repos it's possible to exercise the navigation panel.

Note that the links in the navigation panel are still hard-coded to
mozilla-central/gecko-dev which is a separate pre-existing issue (happens
on e.g. the nss repo on searchfox.org as well).