Follow-up to #30
This changes the type of the `commit` argument to be consistent with the
type of `blame_commit`. Conceptually I believe this is more in line with
the intent of the code, which is to have the function take an optional
commit, but without moving ownership of that thing into the function.