Send a failure email immediately upon failure
Instead of waiting 6 hours for the crontab thing to fire, we can detect
if the indexing exits with an error code, and send an email/shut down
the indexer right away. This patch does so by adding a main.sh wrapper
around index.sh.

Note that as this modifies trigger_indexer.py the lambda task will need
updating.