Tweak completion emails
For dev channel indexing runs:
- Add an email notification when indexing completes
- Send any emails to the author of the HEAD commit instead of the
  searchfox-aws email alias.

This is mostly to improve ergonomics when working on the indexer and
doing dev pushes. The notification email on successful completion is
useful because otherwise it's easy to forget about the indexing run. And
sending the email to the author of the commit is better than spamming
the whole list.