Remove spurious load-vars.sh goop and FILTER variable
Each sub-script that does a particular step of the index building
runs the load-vars script as though it can be run in a standalone
manner. But it can't. There are hidden dependencies on other
variables (most notably the $WORKING variable which is only set
in the indexer-* scripts but used as part of various config
variables like $OBJDIR). So this standalone mode is not a real
thing and we should stop pretending that it is.

This patch does a bunch of cleanup on the load-vars.sh script
and places that use it.

As a bonus it also removes the FILTER variable from some scripts
which seems to be similarly unused and broken.