Bug 1440183 - Fix tokenizing of backtick-strings that contain lines ending in backslash
Backtick strings can have backslashes preceding a newline. Normally newlines
will cause the string token to get split but because of the escaping that the
backslash triggers we end up escaping the newline and so it doesn't split the
token. The equivalent code in the triple-quoted string tokenization doesn't
have this bug because it checks explicitly to see if there is a newline
following the backslash. So I just copied that check over to the backtick
string tokenization and it fixes the problem.