Bug 1451373 - Ensure we index the mozilla subrepo inside c-c as well.
The livegrep indexing is git-aware, and works on a per-repo basis for git repos.
When we switched c-c from using a hg repo to a git repo, we also started going
down the "git repo" codepath in build-codesearch.py, which ignored the mozilla/
subfolder since it wasn't part of the repo. This patch adds the mozilla/
subfolder as another repo to the livegrep config, and ensures it gets searched
along with comm-central. Some fiddling is needed to make sure we adjust the paths
properly to account for the mozilla/ subfolder.

Perhaps in the future we can find a way to reuse the top-level gecko-dev repo
here to save building the livegrep index twice, but that seems more involved and
this gets the job done for now.