Bug 1469956 - Make is_test googlier
There are a bunch of imported gtests that have paths that aren't matched by the existing is_test(). 'gtest' is the generic name given to gtest directories by Mozilla code. 'gtests' is needed in addition to 'gtest' because NSS has a lot of gtests with 'gtests' in the path, but not 'gtest'. 'googletest' is needed because libcubeb and libvpx have that in some of the test paths without 'gtest' or anything else. 'google_test' is used by NSS, but those files are all nested under a 'gtests' directory, so that is not needed.