Allow using an envvar to make the output files diff easily
If you generate two index dirs with different versions of the mozsearch code,
doing a recursive diff on the two file/ directories turns up a lot of spew
because of the changed timestamp. With this patch you can set the
MOZSEARCH_DIFFABLE=1 envvar while running indexer-run.sh to omit the timestamp
and get a cleaner diff. This makes it easier to test changes.