Bug 1466404 - Improve codesearch daemon startup
With the new gRPC mechanism it's easier to get into a state where multiple
queries in quick succession on a fresh server will start the codesearch
daemon multiple times for a single repo. Only one of these will be responding
to queries, but the others sit around and eat memory. (This happens because
the second query arrives before the first query has finished starting the
daemon, and so the second query also tries to start the daemon).

To avoid this, this patch has two robustification steps: (1) start the
codesearch daemon when starting router.py, instead of doing it lazily
on the first query. This ensures the daemon is already up so we don't
attempt to start it multiple times if we get multiple queries in quick
succession. We can still restart it lazily if it dies for some reason, but
in the usual case this avoids the double-start problem entirely.
(2) Poll the webserver less frequently from the indexer when checking to
see if it is up and running. This avoids the problematic scenario for the m-c
repo entirely as it is this rapid polling that is triggers the double start.
Polling every second is unnecessary (and also floods the logs with spurious
failures), increasing the poll interval to 10 seconds works fine.