Ensure we provision rustup.
Previously we were installing rust using static.rust-lang.org which
installs rustc into /usr/local/bin and doesn't install rustup. However,
in order to get the rust stdlib source, it's better to use the more
traditional rustup installation mechanism, which installs it into the
user home folder and also installs rustup. This patch does that and
then uses `rustup update` to update rust as needed, instead of rerunning
the installer in update.sh

This also means we need to run the update.sh script as the ubuntu user,
instead of root, so that rustup is on the $PATH. This is desirable anyway
because running things as root sucks, and the update.sh scripts already
use sudo for individual commands that need root, so in this case running
the script as root is completely avoidable.