Rewrite merge-analyses script in Rust.
This rewrites the merge-analyses.py script in Rust. There is
a minor semantic difference in the Rust version in that it
automatically deduplicates the target lines, while the Python
version does not. This makes it unnecessary for the calling script
to pipe the output through `| sort | uniq` to deduplicate that.
Logically this deduplication step belongs with the merging
tool since the duplicate lines should only appear because of
the need to handle multiple input files.

The output from the Rust version is also better normalized to
a consistent format (with respect to ordering of properties,
etc. within a given line). While the Python version preserved
the order of properties from the input files, the Rust version
takes it a step further and normalizes them to a known order.