Add a read_analyses that slurps multiple files.
This modifies the existing read_analysis function to take a list
of files instead of just a single file, and read the analysis data
from all of those files. It renames the function read_analyses to
be consistent with the change. It adds back a read_analysis function
with the old signature that delegates to the multiple-file version.

When viewing this patch, view the ignore-whitespace diff as most of
the patch is just a change in indentation.