Normalize the analysis data structures.
The analysis data structures are modified slightly to bring them more
in line with what the C++ indexer emits and the documented format of
the analysis data. This consists of a few changes.

The first is a trivial change that reorders some of the fields in the
"source" analysis struct to match that of the C++ indexer. This doesn't
actually matter but I'm OCD like that.

The next is to use a vec of strings to represent the comma-separated
list of symbols allowed in source analysis lines, rather than keeping
it as a single string. Make this change exposed a defect in the output
file formatter where a token may have gotten co-highlighted improperly
because it was using a list of symbols instead of a single symbol,
and so the change to format.rs should fix that by picking a single
symbol.

The final change is to sort and deduplicate the `syntax` and `sym`
fields in the source analysis files, since conceptually these properties
are sets rather than lists. Sorting them provides additional benefits
in terms of producing a normalized output format as the next patch
does.