Impl fmt::Display for the analysis structures.
This implements the Display trait on some structures so that
they can easily be JSON-ified, with the output format being
normalized to what the C++ indexer emits. "Normalized" means
given any set of equivalent input analysis lines, the output
should always be the same - ordering of properties in the
emitted JSON, ordering of tokens with property values that are
comma-separated, omission of optional properties, formatting of
numbers, etc. are all made consistent. This allows us to do
cheap deduplication of equivalent lines, as well as just
being easier to debug.