Bug 1498024 - Further filter the matching tokens by textContent.
The findReferences code returns any tokens anywhere on the patch that
share at least one symbol with the target token. This can sometimes
result in a superset of what we actually want, because "hidden" symbols
(ones that the user doesn't necessarily care about but that get
attributed to tokens) can result in spurious matches.

This patch augments the filtering to ensure only tokens with the same
user-visible string get matched. This is not a perfect fix since it's
still possible to have tokens with the same user-visible strings and
symbols in separate scopes (e.g. different functions) that will get
co-highlighted. But it helps with some cases.