Bug 1507923 - Add caching of the diff tree.
This adds another level of caching, this time across files. If
multiple files are touched in a particular "ignored changeset", then
we would do this N^2 operation because for each of the N files we
would compute a revision-vs-parent diff (which is O(N)). Now, we
just compute the diff the first time and reuse the relevant bits of
the result for future files.