Bug 1515107 - Advance the cur_datum across columns as well
This seems to be a silly error where we don't advance cur_datum properly.
We advance it to the correct line but fail to advance it over to the
correct column. So in some cases we don't find the analysis data for the
token, even though it is present in the dataset.