Bug 1530690 - Don't filter livegrep query by repo
We run different livegrep instances for each entry in the
config.json anyway, so there's no value in doing this repo filtering.
Instead just search all the repos that are part of the target
livegrep instance. This will make it easier to work with submodules
within a repo, since we won't need to list every single one explicitly.