Bug 1282123 - Require specifying the config input file instead of assuming config.json
Currently generate-config.sh takes as input a config file whose name is
assumed to be config.json, and emits another config.json into the working
dir. This patch makes it so that the input config file name is explicitly
provided. This allows us to have different configs inside the
mozsearch-mozilla repo and have them share code, but still deploy
independently.

99% of this patch is just plumbing the new parameter around to various
places. The actual place it gets used is in the `generate-config.sh` script,
where we now use the user-provided config filename instead of a hardcoded
`config.json`.