Bug 1282123 - Add a mechanism for the blame generator to operate on other branches.
This patch adds a BLAME_REF optional environment variable
to transform-repo.py. If this environment variable is not set,
the script has exactly the same effect as before, which is to
read all the commits in the blame repo starting from HEAD, and
then add new commits corresponding to everything reachable
from HEAD in the source repo.

If the BLAME_REF envvar is set, that ref will be used instead
of HEAD for both the source and destination repos. Generally
this should be set to something like 'refs/heads/branchname'
to generate blame commits on the 'branchname' branch from
commits on the 'branchname' branch in the source repo. The
exact semantics are described in a comment in the file.