Bug 1541199 - Serve permalink URLs that point to the head rev directly from nginx.
This patch adds an optimization to the nginx config so that permalink URLs where
the revision matches the repo's HEAD will get served the content directly from
the filesystem, instead of delegating to the rust web-server. This speeds up a
common usecase where permalinks are reloaded on the same searchfox index where
they were generated from the HEAD revision.