Bug 1533802 - Part 2: Impl sticky nesting for C++, rust, JS r=kats r=emilio
This implements position: sticky display of lines for "source" analysis lines
for which we have a "nestingRange" available.  Support for emitting
"nestingRange" is added to the C++ (clang), Rust, and JS analyzer logic.

However, the Rust analyzer changes are very limited because the "save-analysis"
data which forms the basis of the Rust analysis is extremely limited.  It
effectively already is being provided with the result of an analysis pass which
has had all AST information stripped away.  There did not seem to be any
readily available mechanism in the "RLS" to attempt to reconstitute an AST and
relate it to the analysis data.  The best hack I was able to come up with was
to attempt to recursively build up a span covering all of the 'def' children of
a 'def' or an 'impl'.  This means that nesting for rust is largely limited to
"struct" definitions and for inherent "impl" blocks, with the nestingRange only
covering up until the start of the last "fn" in the "impl".