Bug 1550236 - Upgrade clang in vagrant build to clang 8. r=kats
Upgrades to clang-8.  There have been some packaging name changes, with
places where "6.0" was previously used now just being "8".  I added a variable
for the version, but if the naming scheme changes again, that doesn't really
save any effort.  I bumped the alternatives priority so that if someone runs
this on an already-provisioned vagrant, they should get the right thing.

The `add-apt-repository ppa:git-core/ppa` was prompting me when I manually
ran the provisioner script, so I added a `-y`.  (Presumably when run
non-interactively without a pty the prompt wasn't happening, or the fact that
I was running this on an already-provisioned vagrant.)

I also forced the clang-plugin Makefile to use "clang++" since my vagrant
instance was defaulting to g++ (which was really g++, not clang++) if I just
typed make.  It was also necessary to add "-fPIC" because of changes made to
llvm-config made in https://reviews.llvm.org/D55391 as part of
https://bugs.llvm.org/show_bug.cgi?id=8220 that landed in Dec, 2018.