Relax the nesting range's start constraint.
When I refactored the nesting_range to point at the braces, I forgot to update
the formatting logic that decides when to push a symbol onto the nesting stack.
As a result, symbols would only be pushed when the opening brace was on the
same line as the symbol.  Now we only require that the opening brace doesn't
happen before the symbol.