Stop having the search box be position:fixed, instead use flexbox to fix it.
This allows us to stop manually offsetting the "position: sticky" lines to
compensate for the header, which was not reliable using px (and I didn't want
to get into things like rem/etc. or dynamic sizing using JS).

The only real downside to this is that the scrollbar no longer goes to the top
of the content area, but I'm not sure that's a downside.