Bug 1558589 - better nginx caching directives. r=kats
nginx is able to provide us with etags and last-modified times for our static
source files which allows for efficient revalidation, but everything else we
serve is dynamic and currently isn't etag aware.  Because of this, we need
some time-based expiration for at least dynamic stuff.  I've arbitrarily
chosen 2 minutes, but this is motivated primarily by the development cycle.

That said, it's unlikely this will pose a major problem.