[wip] Move to Ubuntu 18.04 LTS.
And also allow libvirt as a back end, so I don't have to deal with incompatible
VirtualBox drivers :(

I _think_ the main issue here is that the disk space goes down to the regular
one (40GB), which may be not enough for mozilla-index...

Testing that with the indexing atm. But test repo and provisioning works fine.

Also I guess we'd need to update the AMIs.