Make NFS settings libvirt-specific
Although our intent when adding libvirt was to add the NFS mount only in the
case of libvirt and not for virtualbox, we did not, so it was also being
added for virtualbox on OS X which did not work.  This corrects the syntax to
only enable NFS on libvirt.

This also has a ride-along of updating to the most recent released image
version, primarily because I tried this as part of my troubleshooting before
realizing what was going on, but it seems like an advantageous change.