Import servo's malloc_size_of in order to fork it.
We run into the awkward problem where wr_malloc_size_of is missing some
useful things like support for BTreeMap and the only place the impl for
BTreeMap can be implemented is in the wr_malloc_size_of crate.  So then it
becomes a question of which malloc_size_of implementation to fork.

And unfortunately we do need to fork it, because malloc_size_of has
servo-internal dependencies like `hashglobe`, its forked version of
std::HashMap.