Bug 1611729 - Rewrite and simplify blame popup.
With the goal of making it faster.

Right now the blame popup causes pretty expensive reflows, because it's
hosted inside the line strip.

This is completely unnecessary though, so this patch saves us some
per-line markup and makes it a top-level absolutely-positioned element,
and positions it dynamically.

This also uses aria-owns so that the accessibility semantics are the
same as before.