No bug - Rewrite the code highlighter.
This is another piece of the code which was doing something relatively
simple, in an extremely complex way.

The only thing this needs to do is keeping a set of selected lines and
represent them in the DOM, knowing how to highlight them, serializing
them from an URL hash, and parsing that URL hash.

I couldn't understand any of the previous code. We don't use the
"clicked" and "multihightlight" classes, so unify that code.

This still uses createSyntheticAnchor. I'll get to that.