No bug - Move synthetic anchor code to code-highlighter.js
The first call was not needed anyway because the first thing the
highlighter did was calling updateFromHash, which ends up in updateHash
(if there's any valid hash) and thus in createSyntheticAnchor.

The second call is moved to the constructor. The navigate thing is not
needed in Chromium stable so I just ripped it off.