Bug 1627532 - Use cinnabar unless explicitly disabled.
Instead of explicitly requiring CINNABAR=1 to enable cinnabar
lookups, make that the default and allow CINNABAR=0 to disable
cinnabar lookups. In general the cinnabar lookup is pretty cheap
and on repos that don't have cinnabar data get_hg_rev will return
None all the time so we won't put any hg rev in the commit
message. The main purpose of this change is to simplify the call
sites for the common case.