No bug - Rewrite most of the dxr / AJAX logic.
This gets rid of most of the jQuery usage in the tree, and uses fetch()
instead.

I've left most of the result rendering code and such completely as-is
for now (just de-indented), since this was a big enough patch already.

The only behavior change this has is that the info bubbles and such
shouldn't fade in/out. I think it's not a big deal, but let me know if
you disagree.