Bug 1636498 - Remove UA sniffing hack.
Use the same approach to display line numbers in all browsers.

There were two conflicting things that caused the current state of
affairs:

 * We wanted line numbers not to be selectable, so that copying text
   works.

 * We wanted line numbers to be searchable.

We achieved these on Firefox via regular text and user-select: none.

On other browsers we couldn't achieve both, because they don't allow to
search pseudo-elements, and user-select actually doesn't quite work as
it should.

However! I recently fixed a bug in Firefox[1] that makes us able to
find-in-page pseudo-elements!

So we'd switch to both benefits in Firefox Nightly and Beta, but the
same behavior everywhere else, for free.

[1]: https://bugzilla.mozilla.org/show_bug.cgi?id=1627643