Bug 1616362 - Migrate the user-facing AWS scripts from py2 to py3
This allows developers to create their local virtualenv with python3
and everything should still work. The terminate-indexer.py script is
also run from the indexer itself, and that works fine with py3. The
trigger_indexer.py script is also run from the lambda task, which is
also migrated to py3 here.