Bug 1616362 - Run router.py with py3.
This includes a handful of changes that are needed when running with python3,
mostly around the fact that python3 distinguishes between bytes and strings
better, so we need to convert back and forth in a few places.

Note that this patch makes the router not compatible with python2, but if
we need to revert back to python2 we just need to back out this one patch.