Bug 1640915 - Report stray index volumes.
The only time there should be unattached index volumes are (a) immediately
after creation and before attaching to the indexer, and (b) while detaching
from the indexer and reattaching to the web-server. These are both very short
periods of time, so if we find stray index volumes after deleting old web-servers
that's likely a problem. Report these as warnings for further investigation.