Bug 1616362 - Rewrite find-objdir-files.py in bash
This also drops the "conftest" special-casing which doesn't seem to be needed
any more, as there are no files/directories containing that substring in the
current m-c analysis/__GENERATED__ folder. I had initially retained that by
inserting a `| grep -v conftest` between the `find` and the `sed` commands
but it produced a nonzero exit code because it wouldn't find anything. Instead
of working around that it seemed simpler to just ditch it as it doesn't seem
needed.