Bug 1646137 - Stop provisioning Firefox build stuff that we don't need anymore.
We still need pkg-config so explicitly install that. The web-server script
already does this but it was missing in the indexer, because the Firefox
build stuff pulled it in as a dependency.

We also need zlib.h on the indexer for NSS, and mercurial for cinnabar,
so explicitly install those too.