Bug 1653245 - Use mempack to write packfiles instead of loose objects.
Instead of writing the blame commits into the blame repo as loose objects,
which take up a lot of disk space, they now get written into a mempack backend
first. The mempack backend is an in-memory store that can be flushed directly
into a packfile which is much more space-efficient, specially during initial
blame repo creation.