Bug 1653986 - make info_boxes flow slightly more idiomatic. r=kats
The InfoBox structure directly contains some `F` formatting and in
`generate_info_boxes` we build a derived F formatting via
`F::Indent(info_box.body_nodes)` which previously moved the formatting.
This led to the argument passing moving the vector out of necessity (and
failing to provide a comment to this effect), but it's more idiomatic to
provide a borrowed slice reference.

Because the F enum does not have a borrowing mode of operation, I added
a `#[derive(Clone)]` and changed the code to
`F::Indent(info_box.body_nodes.clone())`.