Bug 1631466 - enhance gzipping to run in parallel. r=kats
The previous approach worked but didn't leverage the multiple CPU cores
available on the indexer.   It took 17 minutes for mozilla-central to
compress, which justifies dealing with parallelism.

We handle the .gz suffix case by splitting the compression into 2
passes.  This adds slight complexity because grep normally likes to
complain if it finds no matches and there's no guarantee we'll have a
.gz match.