Bug 1417127 - Fix some edge cases in the tokenizer
peek_char() returns a sentinel character of '!' when the end of
the string is reached. So doing things like:
  if peek_char() != 'x' {
    get_char();
  }
can trigger a call to get_char() if we're at the end of the string,
and that causes get_char() to panic. This was uncovered by the
clang/test/Lexer/eof-char.c testcase in the llvm repository,
which exercises one such code path in the tokenizer.

This patch introduces a peek_isnot function meant to be used in
these cases and which explicitly returns false when at the end of
the string. The patch uses the function at places that are
susceptible to this bug.