Bug 1630646 - Disable port reuse in codesearch.
This has the side-effect that if multiple codesearch instances are
started on the same port, only the first one will succeed and the
rest will fail early. (Previously they would all succeed and the
OS would load-balance incoming requests between the multiple instances).
This ensures that if the codesearch instance goes down while router.py
is handling a flood of requests, it doesn't end up starting a ton
of new codesearch instances that then OOM the machine. Instead,
only one will start up, and some of the incoming requests may get
temporary server errors until it's back up and running.