Bug 1670314 - Followup for release channel rename
I forgot that we have the release channel hard-coded in one more spot,
where we decide to terminate the old web-server. We have code to
special-case the release (now release1) channel so that we can roll
back to a day-old web server for m-c in case of emergency. This patch
updates the channel name there so we preserve the old behaviour.