Bug 1671640 - Ensure stderr is properly captured
Based on the code later in this function, it seems clear that the intent
is to capture stderr into the `stderr` variable. However that's not
actually happening. This patch corrects that and makes it happen.

A beneficial side-effect is that the codesearch warnings about ignoring
too-big files now don't end up in the indexer log and so won't trigger
warning emails.