Bug 1675270 - Drop filesystem checks from web-server-check.sh
If this is to be run in production, it will run after the web-server
has been started. In that case the local filesystem checks can't happen
because the volume with those files has been transferred to the web-server.
The local filesystem check will still run as part of mkindex.sh on the
indexer.