Bug 1674601 - Make the blame/coverage strips a bit more touch friendly
Instead of using the touch-generated mouse events to control the blame popup,
add touch event listeners to manage it directly. This patch allows the user
to drag their finger around on the blame strip, and have the blame popup update
to whatever is under the cursor. The touch-action CSS property is used to
disable browser-based panning while the finger is being dragged in such a
manner. This UX makes it somewhat easier to trigger the blame popup for a
given line using touch, where the finger is less precise of a pointer than
a mouse cursor. When triggered via touch, the blame popup remains visible
until it is dismissed by scrolling or "clicking" (i.e. tapping) elsewhere.