Bug 1654946 - Update build-blame to use git-fast-import.
Instead of using the libgit2 API to build structures in the blame
repository, we now open a git-fast-import subcommand and stream
the contents into that. This approach is faster for large repos,
because the object/packfile management is faster than can be
achieved with the currently exposed libgit2 API.

The build-blame tool spawns worker threads to compute diffs; the
code there is unmodified in this patch. The main thread code that
actually writes to the blame repo is what is being modified here.

As part of generating the new commits, the code needs to know
oids for tree objects that were not changed, and contents of
blobs written in previous commits. The mechanism for doing this
changes in this patch, because they need to be requested from
git-fast-import instead of reading them using the libgit2 API.
In this patch, changes to the helper functions are in service
of accomplishing this.

The other main change in this patch is the main loop that writes
the commit metadata into the repo; that now needs to write
the relevant information to the fast-import stream instead of
using the libgit2 API. Conceptually this change is relatively
straightforward.