Bug 1683563 - Add blame rebuilding scripts
The trigger_blame_rebuild script is almost identical to trigger_indexer.py,
and is intended to be run manually by devs who want to trigger a blame
rebuild. As with indexer runs, it can run using code from a specific branch/fork
and on a specific channel to test changes before actually deploying them.

The entry point on the spawned EC2 instance is main.sh (same as with a regular
indexer) but it then delegates to rebuild-blame.sh instead of index.sh. That
handles some of the AWS setup/teardown steps while delegating the actual
blame rebuilding steps to infrastructure/reblame-run.sh.

This also adds a Makefile target that shows how to exercise the reblame-run.sh
script in local vagrant environment on the NSS repo.

Unlike the indexer, running trigger_blame_rebuild won't spawn a web server
instance. Instead, it just uploads the new blame tarballs to S3 (with the -new
suffix) and terminates itself.