Bug 1694210 - Update AMI image name refs, use python2 for logs
The VM triggering scripts need to know the revised AMI name that we're
going to use.  Also, during provisioning, the awslogs-agent-setup.py
script failed to execute properly because the embedded hash-bang was
for "python", so we now explicitly run it with python2 because our
python3 is too new.

After this I'll also regenerate the lambda jobs.