Bug 1694210 - Explicitly use up-to-date version of certifi
On my machine, the current virtualenv steps resulted in certifi
version 2019.11.28 being installed into the virtualenv as created by
the lambda.zip creation process.  The resulting lambda jobs failed to
authenticate the AWS server.

This patch is a hack in the sense that I want the lambda jobs to be
happy again (and they now are, with the changes in this patch), but I'm
not sure it's the right answer.  My thoughts here were:
1. Validating certificates seems good, why not.
2. Upgrading certifi makes the lambda job work.

By default, virtualenv uses a seeding mechanism to install things beyond
just python / python3 / python 3.N links into the env/bin directory.
This is documented at
https://virtualenv.pypa.io/en/latest/user_guide.html but I haven't
fully processed it in detail.

The short story of my investigation is:
- Passing `--no-seed` leaves us with only the python aliases which is
  explicitly not what we want.
- Just passing "certifi" to install commands is insufficient to force
  a more recent version than was available in the seeding process.
- Explicitly triggering an upgrade gets us the current certifi version.