Bug 1701848 - Add script to aid in analyzing searchfox web-server dynamic query performance
Example of output can be found in the bug.

This depends on https://github.com/rcoh/angle-grinder which can be installed
via `cargo install ag`.  My current assumption is that this is being run
locally, not in a VM, by someone who has already done this.

I'm not sure I have a longer term strategy for this, I just wanted to archive
the results of my investigation in a way that's slightly better than searching
for a bugzilla comment and then copying and pasting from that.