Bug 1414954 - Add secret codesearch support for getting lines of context.
This implements an undocumented "context:N" search directive understood by
the parser (if it precedes the search query proper, as-is required by the
implementation of parse_search) that we propagate to livegrep which in turn
generously returns the context for us.

I'm calling this secret because I don't think the UX as provided by this
patch is appropriate for a supported feature, but I think this could be
useful for experimentation and iteration.  In particular, in this patch:
- Typing out context:N at the start of the search but having things not work
  if you put it at the end as an afterthought isn't great.
- I don't think adding a separate form input for lines of context is a great
  initial choice either.
- This doesn't work with any of the semantic responses, and doing so is both
  non-trivial and not something I think we should be freshly implementing in
  Python.
- On the actual result line where we have bounds, we normalize off leading
  whitespace whereas we leave the whitespace intact on the context lines and
  this looks weird.